refactor: typescript 설치 및 환경설정
타입스크립트 적용 없이 몇 js 파일의 확장자만 ts로 바꿔봄
