+++ package.json
@@ -74,6 +74,7 @@
     "eslint-plugin-storybook": "^0.5.11",
     "gh-pages": "^3.1.0",
     "prettier": "2.6.2",
+    "typescript": "^4.7.2",
     "webpack": "^5.72.0"
   }
 }
